requirements.txt
django==4.2.21
djangorestframework
mongoengine
blinker